Real_time_System/requirements.txt
pandas
pandas-ta
FiinQuantX
streamlit
python-dotenv
requests
schedule
